backend/requirements.txt
crewai
langchain
langchain-google-genai
langchain-experimental
mftool
deprecated
pandas
python-dotenv